+++ core/data/src/main/java/com/startup/data/datasource/AuthDataSource.kt
@@ -3,8 +3,7 @@ package com.startup.data.datasource
 import kotlinx.coroutines.flow.Flow
 
 interface AuthDataSource {
-    fun kakaoLogin(): Flow<String>
-    fun login(): Flow<Unit>
+    fun login(kakaoToken: String): Flow<Unit>
     fun unLink(): Flow<Unit>
     fun logOut(): Flow<Unit>
     fun localLogout(): Flow<Unit>
