+++ core/data/src/main/java/com/startup/data/remote/datasourceimpl/AuthDataSourceImpl.kt
@@ -1,7 +1,6 @@
 package com.startup.data.remote.datasourceimpl
 
 import androidx.datastore.preferences.core.Preferences
-import com.startup.common.util.KakaoAuthFailException
 import com.startup.common.util.ResponseErrorException
 import com.startup.common.util.UserAuthNotFoundException
 import com.startup.data.datasource.AuthDataSource
@@ -10,27 +9,21 @@ import com.startup.data.local.provider.TokenDataStoreProvider
 import com.startup.data.remote.dto.request.auth.JoinRequest
 import com.startup.data.remote.dto.request.auth.LoginRequest
 import com.startup.data.remote.ext.emitRemote
-import com.startup.data.remote.ext.requireNotNull
 import com.startup.data.remote.service.AuthService
 import com.startup.data.remote.service.MemberService
 import com.startup.data.util.ACCESS_TOKEN_KEY_NAME
-import com.startup.data.util.KakaoLoginClient
+import com.startup.data.util.KakaoLogoutHelper
 import com.startup.data.util.REFRESH_ACCESS_TOKEN_KEY_NAME
 import com.startup.data.util.handleExceptionIfNeed
-import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.onEmpty
 import javax.inject.Inject
 import javax.inject.Named
 
 internal class AuthDataSourceImpl @Inject constructor(
-    private val kakaoLoginClient: KakaoLoginClient,
+    private val kakaoLogoutHelper: KakaoLogoutHelper,
     private val authService: AuthService,
     private val memberService: MemberService,
     private val tokenDataStoreProvider: TokenDataStoreProvider,
@@ -38,55 +31,45 @@ internal class AuthDataSourceImpl @Inject constructor(
     @Named(ACCESS_TOKEN_KEY_NAME) private val accessTokenKey: Preferences.Key<String>,
     @Named(REFRESH_ACCESS_TOKEN_KEY_NAME) private val refreshTokenKey: Preferences.Key<String>,
 ) : AuthDataSource {
-    override fun kakaoLogin(): Flow<String> = flow {
-        kakaoLoginClient.login().fold(
-            onSuccess = { token ->
-                emit(token.accessToken)
-            },
-            onFailure = { error ->
-                throw KakaoAuthFailException("로그인 오류 : $error")
-            }
-        )
-    }
 
-    override fun login(): Flow<Unit> = callbackFlow {
-        kakaoLogin().onEmpty {
-            throw KakaoAuthFailException("로그인 오류")
-        }.map {
-            val item = authService.login(LoginRequest(provider = "kakao", loginAccessToken = it))
-            if (item.data != null && item.data.accessToken == null) {
-                throw UserAuthNotFoundException(
-                    message = "계정 없음",
-                    nickName = kakaoLoginClient.me().getOrNull()?.kakaoAccount?.profile?.nickname,
-                    providerToken = it
+    override fun login(kakaoToken: String): Flow<Unit> = flow {
+        emitRemote(
+            authService.login(
+                LoginRequest(
+                    provider = "kakao",
+                    loginAccessToken = kakaoToken
                 )
-            }
-            item.data.requireNotNull()
-        }.onEach {
-            tokenDataStoreProvider.putValue(accessTokenKey, it.accessToken!!)
-            tokenDataStoreProvider.putValue(refreshTokenKey, it.refreshToken!!)
-            send(Unit)
-        }.catch {
-            if (it is UserAuthNotFoundException) {
-                throw it
-            } else {
-                throw ResponseErrorException("로그인 오류")
-            }
-        }.launchIn(this)
-        awaitClose()
+            )
+        )
+    }.map {
+        if (it == null || it.accessToken.isNullOrBlank()) {
+            throw UserAuthNotFoundException(
+                message = "계정 없음",
+                nickName = kakaoLogoutHelper.me().getOrNull()?.kakaoAccount?.profile?.nickname,
+                providerToken = kakaoToken
+            )
+        }
+        tokenDataStoreProvider.putValue(accessTokenKey, it.accessToken)
+        tokenDataStoreProvider.putValue(refreshTokenKey, it.refreshToken!!)
+    }.catch {
+        if (it is UserAuthNotFoundException) {
+            throw it
+        } else {
+            throw ResponseErrorException("로그인 오류")
+        }
     }
 
     override fun unLink(): Flow<Unit> = flow {
         handleExceptionIfNeed {
             emitRemote(memberService.withdrawalService())
-            kakaoLoginClient.unLink()
+            kakaoLogoutHelper.unLink()
             logoutManager.logout()
         }
     }
 
     override fun logOut(): Flow<Unit> = flow {
         handleExceptionIfNeed {
-            kakaoLoginClient.logout()
+            kakaoLogoutHelper.logout()
             logoutManager.logout()
             emitRemote(memberService.logoutService())
         }
@@ -110,7 +93,6 @@ internal class AuthDataSourceImpl @Inject constructor(
     }.map {
         tokenDataStoreProvider.putValue(accessTokenKey, it.accessToken!!)
         tokenDataStoreProvider.putValue(refreshTokenKey, it.refreshToken!!)
-        Unit
     }
 
     override fun getAccessToken(): Flow<String> =
