+++ core/domain/src/main/java/com/startup/domain/repository/AuthRepository.kt
@@ -3,7 +3,7 @@ package com.startup.domain.repository
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
-    fun login(): Flow<Unit>
+    fun login(kakaoToken: String): Flow<Unit>
     fun unLink(): Flow<Unit>
     fun logOut(): Flow<Unit>
     fun localLogout(): Flow<Unit>
