+++ feature/login/build.gradle.kts
@@ -9,4 +9,5 @@ android {
 dependencies {
     implementation(libs.androidx.core.splashscreen)
     implementation(libs.kotlinx.coroutines.android)
+    implementation(libs.kakao.login)
 }
